requirements.txt
appdirs==1.4.4
black==23.3.0
click==8.1.3
colorama==0.4.6
cycler==0.11.0
kiwisolver==1.4.4
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.2
opencv-python
pathspec==0.11.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.3.23
six==1.16.0
toml==0.10.2
pyserial==3.5
PyQt5
qdarkstyle